Update code to use FunctionGet / ClassGet
