+++ modal/object.py
@@ -336,98 +336,8 @@ class _StatefulObject(_Object):
             return False
 
 
-class _AppObject(_Object):
-    # Similar to _StatefulObject, this is a temporary class needed until we rewrite AppLookupObject
-    # to be specific to each class. This base class will only be used for _Function and _Cls
-    # Those are always looked up using an app name and a tag.
-
-    @classmethod
-    def from_name(
-        cls: Type[O],
-        app_name: str,
-        tag: Optional[str] = None,
-        namespace=api_pb2.DEPLOYMENT_NAMESPACE_WORKSPACE,
-        environment_name: Optional[str] = None,
-    ) -> O:
-        """Retrieve a function/class with a given name and tag.
-
-        **Examples**
-
-        ```python notest
-        # Retrieve a function from a different app
-        stub.other_function = Function.from_name("other-app", "function")
-        ```
-        """
-
-        async def _load_remote(obj: _Object, resolver: Resolver, existing_object_id: Optional[str]):
-            request = api_pb2.AppLookupObjectRequest(
-                app_name=app_name,
-                object_tag=tag,
-                namespace=namespace,
-                object_entity=cls._type_prefix,
-                environment_name=_get_environment_name(environment_name, resolver),
-            )
-            try:
-                response = await retry_transient_errors(resolver.client.stub.AppLookupObject, request)
-            except GRPCError as exc:
-                if exc.status == Status.NOT_FOUND:
-                    raise NotFoundError(exc.message)
-                else:
-                    raise
-
-            handle_metadata = get_proto_oneof(response.object, "handle_metadata_oneof")
-            obj._hydrate(response.object.object_id, resolver.client, handle_metadata)
-
-        rep = f"Ref({app_name})"
-        return cls._from_loader(_load_remote, rep, is_another_app=True)
-
-    @classmethod
-    async def lookup(
-        cls: Type[O],
-        app_name: str,
-        tag: Optional[str] = None,
-        namespace=api_pb2.DEPLOYMENT_NAMESPACE_WORKSPACE,
-        client: Optional[_Client] = None,
-        environment_name: Optional[str] = None,
-    ) -> O:
-        """Lookup a function/class with a given name and tag.
-
-        Example: Look up a function from a different app:
-
-        ```python notest
-        other_function = Function.lookup("other-app", "function")
-        ```
-        """
-        obj = cls.from_name(app_name, tag, namespace=namespace, environment_name=environment_name)
-        if client is None:
-            client = await _Client.from_env()
-        resolver = Resolver(client=client)
-        await resolver.load(obj)
-        return obj
-
-    @classmethod
-    async def _exists(
-        cls: Type[O],
-        app_name: str,
-        tag: Optional[str] = None,
-        namespace=api_pb2.DEPLOYMENT_NAMESPACE_WORKSPACE,
-        client: Optional[_Client] = None,
-        environment_name: Optional[str] = None,
-    ) -> bool:
-        """Internal for now - will make this "public" later."""
-        try:
-            await cls.lookup(app_name, tag, namespace, client, environment_name)
-            return True
-        except GRPCError as exc:
-            if exc.status == Status.NOT_FOUND:
-                return False
-            else:
-                raise
-
-
 Object = synchronize_api(_Object, target_module=__name__)
 StatefulObject = synchronize_api(_StatefulObject, target_module=__name__)
-AppObject = synchronize_api(_AppObject, target_module=__name__)
 
 
 def live_method(method):
